Fix float64 casting.

+++ models/staging/stg_detalhamento_pedido.sql
@@ -4,8 +4,8 @@ with
             cast(salesorderid as int) as ID_VENDA
             , cast(productid as int) as ID_PRODUTO
             , cast(orderqty as int) as QUANTIDADE_COMPRADA
-            , cast(unitprice as float) as PRECO_UNITARIO
-            , cast(unitpricediscount as float) as DESCONTO_PERCENTUAL_UNITARIO
+            , cast(unitprice as float64) as PRECO_UNITARIO
+            , cast(unitpricediscount as float64) as DESCONTO_PERCENTUAL_UNITARIO
         from {{source('indicium-ae-certification','salesorderdetail')}}
 )
